==> susquehanna/notebooks/12. Feature scoring analysis.ipynb <==
import warnings

warnings.simplefilter(action="ignore", category=FutureWarning)
import sys
import matplotlib.pyplot as plt
import numpy as np
import pandas as pd
import seaborn as sns
#import ipyparallel
from ema_workbench.analysis import feature_scoring
import os

sys.path.append("..")
import rbf_functions

rbfs = [
    rbf_functions.original_rbf,
    rbf_functions.squared_exponential_rbf,
    rbf_functions.inverse_quadratic_rbf,
    rbf_functions.inverse_multiquadric_rbf,
    rbf_functions.exponential_rbf,
    rbf_functions.matern32_rbf,
    rbf_functions.matern52_rbf,
]

refset = {}
varset = {}
for entry in rbfs:
    name = entry.__name__
    # load variables
    output_dir = os.path.abspath("../notebooks/refsets/")
    for filename in os.listdir(output_dir):
        if filename == f"{name}_refset.csv":
            df_refs = pd.read_csv(f"{output_dir}/{filename}")
            refset[name] = df_refs
        elif filename == f"{name}_refset_variables.csv":
        #elif filename == f"{name}_refset_with_variables.csv":
            df_vars = pd.read_csv(f"{output_dir}/{filename}", header=None)
            columnname = [
                "c1",
                "r1",
                "c2",
                "r2",
                "c3",
                "r3",
                "c4",
                "r4",
                "c5",
                "r5",
                "c6",
                "r6",
                "c7",
                "r7",
                "c8",
                "r8",
                "w01",
                "w02",
                "w03",
                "w04",
                "w05",
                "w06",
                "w07",
                "w08",
                "w09",
                "w10",
                "w11",
                "w12",
                "w13",
                "w14",
                "w15",
                "w16",
            ]
            df_vars.columns = columnname
            varset[name] = df_vars
varset.keys()

refset["original_rbf"].head()

varset["original_rbf"].head()

#### sns.set(rc={'figure.figsize':(11.7, 8.27)})

for i in refset:
    #     print(i)
    x = varset[i]
    y = {k: np.asarray(v) for k, v in refset[i].to_dict("list").items()}

    scores = feature_scoring.get_feature_scores_all(x, y, alg="extra trees")
    g = sns.heatmap(scores, annot=True, cmap="viridis")
    #     plt.show()
    plt.tight_layout()
    #     plt.savefig(f"feature_score/{i}_fsc")
    #     plt.close()
    plt.show()
